% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lambda.R
\name{create_api_integration}
\alias{create_api_integration}
\title{Create an API integration}
\usage{
create_api_integration(
  aws_profile,
  api_name,
  resource_name,
  aws_region,
  request_parameter,
  lambda_arn,
  deployment_stage,
  package = "awslambdafuns"
)
}
\arguments{
\item{package}{}
}
\description{
Create an API integration
}
