% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lambda.R
\name{create_lambda_fun}
\alias{create_lambda_fun}
\title{Create an AWS Lambda function}
\usage{
create_lambda_fun(
  aws_profile = "default",
  function_name,
  aws_region = "eu-west-2",
  aws_account_id = Sys.getenv("AWS_ACCOUNT_ID"),
  aws_service_role = Sys.getenv("AWS_SERVICE_ROLE"),
  memory_size = 1500,
  skip = FALSE,
  package = "awslambdafuns"
)
}
\arguments{
\item{package}{}
}
\description{
Create an AWS Lambda function
}
