% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lambda.R
\name{do_lambda_fun}
\alias{do_lambda_fun}
\title{A low-level function to do something in AWS Lambda}
\usage{
do_lambda_fun(
  script,
  aws_profile = "default",
  function_name,
  aws_region,
  aws_account_id = Sys.getenv("AWS_ACCOUNT_ID"),
  aws_service_role = Sys.getenv("AWS_SERVICE_ROLE"),
  memory_size = 1500,
  skip = FALSE,
  package = "awslambdafuns"
)
}
\arguments{
\item{function_name}{Should match a folder placed in inst/lambda_funs}

\item{package}{}
}
\description{
A low-level function to do something in AWS Lambda
}
