% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lambda.R
\name{update_lambda_fun}
\alias{update_lambda_fun}
\title{Update an AWS Lambda function}
\usage{
update_lambda_fun(
  aws_profile,
  function_name,
  aws_region = "eu-west-2",
  account_id,
  skip = FALSE,
  package = "awslambdafuns"
)
}
\arguments{
\item{package}{}
}
\description{
Update an AWS Lambda function
}
